Move and rename AbsenceType to CalendarAbsenceType
