--- src/main/java/org/synyx/urlaubsverwaltung/absence/AbsenceType.java
@@ -1,7 +1,0 @@
-package org.synyx.urlaubsverwaltung.absence;
-
-public enum AbsenceType {
-
-    DEFAULT,
-    HOLIDAY_REPLACEMENT
-}
