+++ src/main/java/org/synyx/urlaubsverwaltung/calendar/CalendarAbsenceType.java
@@ -1,0 +1,7 @@
+package org.synyx.urlaubsverwaltung.calendar;
+
+public enum CalendarAbsenceType {
+
+    DEFAULT,
+    HOLIDAY_REPLACEMENT
+}
